traefik: fix ingress status
When running behind a service of type loadbalancer, traefik can copy the
this service's status onto the ingress to show whether the ingress
controller is ready to serve traffic. Sadly, this doesn't work for
services of type ClusterIP or NodePort, since their status field will
remain empty. In this case ArgoCD will view the ingress as continously
progressing and never healthy.

This commit adds an additional command to the traefik pod which sets a
canned hostname=nodeport response on all ingresses controlled by
traefik, which will satisfy ArgoCD or other controllers watching the
status