jsonnet: add k8s-libsonnet@v1.21 and make it available as k.libsonnet
Grafana has created a very useful jsonnet library which can be used to
generate all kubernetes objects. In order to allow for easy upgrading of
the target k8s version and to facilitate compatibility with older
ksonnet style libraries, a simple indirection in the form of k.libsonnet
is provided in /lib. While ksonnet and tanka style projects would expect
'k.libsonnet' to be on the JPath directly, all files in this repo should
import k.libsonnet via relative path to avoid needing to propagate the
JPath to sub-apps